Minor refactor of simulation
