+++ ui/src/components/branch-graph/BranchNode.tsx
@@ -14,8 +14,8 @@ export function BranchNode({
 	onClick?: () => void;
 }) {
 	const transform = useComputedAtom((get) => {
-		const { x, y } = get(node.atom);
-		return `translate(${(x ?? 0).toFixed(1)}px, ${(y ?? 0).toFixed(1)}px)`;
+		const { screenX, screenY } = get(node.atom);
+		return `translate(${screenX.toFixed(1)}px, ${screenY.toFixed(1)}px)`;
 	});
 	return (
 		<JotaiG style={{ transform: transform }} {...drag(node, onMove, onClick)}>
